requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
MouseInfo==0.1.3
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.0
psutil==5.8.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.26
python-telegram-bot==13.3
PyTweening==1.0.3
pytz==2021.1
requests==2.25.1
six==1.15.0
speedtest-cli==2.1.2
telebot==0.0.4
telegram==0.0.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.3
